kishan/requirements.txt
Django==3.2.8 
djangorestframework==3.12.4
Pillow==8.3.2